Challenge-1A/Adobe/requirements.txt
pdfplumber==0.10.2
PyMuPDF==1.23.6
scikit-learn==1.2.2
numpy==1.23.5
joblib==1.3.2
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.2.0
